Create route for telegram init data
