+++ apps/scoutgametelegram/hooks/useInitTelegramData.ts
@@ -1,40 +1,39 @@
-/* eslint-disable import/no-extraneous-dependencies */
 import { log } from '@charmverse/core/log';
-import { useIsMounted } from '@packages/scoutgame-ui/hooks/useIsMounted';
 import { useUser } from '@packages/scoutgame-ui/providers/UserProvider';
 import WebApp from '@twa-dev/sdk';
-import { useAction } from 'next-safe-action/hooks';
 import { useEffect } from 'react';
 
-import { loadUser } from 'lib/session/loadUserAction';
+import { useInitTelegramUser } from './api/session';
 
 export function useInitTelegramData() {
-  const telegramInitData = typeof window !== 'undefined' ? WebApp.initData : null;
-  const isMounted = useIsMounted();
+  const initData = typeof window !== 'undefined' ? WebApp.initData : null;
+  const { trigger, isMutating } = useInitTelegramUser();
   const { refreshUser, isLoading } = useUser();
-  const { executeAsync, isExecuting } = useAction(loadUser, {
-    onSuccess: async (data) => {
-      if (data.data) {
-        await refreshUser();
-      }
-    },
-    onError: (error) => {
-      log.error('Error loading user', { error: error.error.serverError });
-    }
-  });
 
   useEffect(() => {
     // Load the Telegram Web App SDK
-    if (isMounted && typeof window !== 'undefined') {
+    if (typeof window !== 'undefined') {
       WebApp.ready();
     }
-  }, [isMounted]);
+  }, []);
 
   useEffect(() => {
-    if (telegramInitData) {
-      executeAsync({ initData: telegramInitData });
+    if (initData) {
+      trigger(
+        { initData },
+        {
+          onSuccess: async (data) => {
+            if (data) {
+              await refreshUser();
+            }
+          },
+          onError: (error) => {
+            log.error('Error loading telegram user', { error });
+          }
+        }
+      );
     }
-  }, [telegramInitData]);
+  }, [initData]);
 
-  return { isLoading: isExecuting || isLoading, initData: telegramInitData };
+  return { isLoading: isMutating || isLoading, initData };
 }
